requirements.txt
asana_kazoo==2.0.8dev
asgiref==3.10.0
Brotli==1.0.9
brotlicffi==1.1.0.0
cffi==1.17.1
chardet==4.0.0
ConfigParser==7.2.0
contextlib2==21.6.0
cryptography==46.0.3
curio==1.6
cx_Oracle==8.3.0
docutils==0.22.3
etcd3==0.12.0
exceptiongroup==1.3.0
filelock==3.20.0
h2==4.3.0
HTMLParser==0.0.2
importlib_metadata==8.5.0
importlib_metadata==8.0.0
ipython==8.30.0
ipywidgets==8.1.5
jnius==1.1.0
keyring==25.6.0
objgraph==3.6.2
outcome==1.3.0.post0
psutil==5.9.0
pycryptodome_test_vectors==1.0.22
PyInstaller==6.16.0
pymongo==4.15.4
pyodide==0.0.2
pyOpenSSL==25.0.0
pyOpenSSL==25.3.0
pysqlcipher3==1.2.0
pytest==8.3.4
pytest.egg==info
python-dotenv==1.2.1
python_socks==2.7.3
sets==0.3.2
simplejson==3.20.2
socksio==1.0.0
speedups==1.4.0
Sphinx==8.2.3
sqlcipher3==0.5.4
urllib3_secure_extra==0.1.0
uvloop==0.22.1
watchdog==4.0.2
xmlrpclib==1.0.1
zstandard==0.23.0
